# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Politico is an app that helps both the politicians and citizens. It helps citizens give mandates to politicians running for different offices",
    "version": "1.0.0",
    "title": "Politico",
    "termsOfService": "",
    "contact": {
      "email": "samabos007@gmail.com"
    }
  },
  "host": "samson-politico.herokuapp.com",
  "basePath": "/api/v1",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/parties/{partyid}": {
      "get": {
        "tags": [
          "Political Parties"
        ],
        "summary": "displays specific political party by id",
        "description": "",
        "operationId": "partyid",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "partyid",
            "description": "ID of political party to display",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/parties"
            }
          },
          "404": {
            "description": "The record with the given id was not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Political Parties"
        ],
        "summary": "Delete political party by ID",
        "description": "A valid uuid type of the party to delete must be supplied",
        "operationId": "deleteParty",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "partyid",
            "in": "path",
            "description": "ID of the party that needs to be deleted",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Party Delete Successful"
          },
          "404": {
            "description": "Party not found"
          }
        }
      }
    },
...
    "/office/{officeid}/": {
      "get": {
        "tags": [
          "Political Offices"
        ],
        "summary": "displays specific political office by id",
        "description": "",
        "operationId": "officeid",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "officeid",
            "description": "ID of political office to display",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/offices"
            }
          },
          "404": {
            "description": "The record with the given id was not found"
          }
        }
      }
    },
...
    "/vote/": {
      "post": {
        "tags": [
          "Vote"
        ],
        "summary": "Vote for an Office",
        "description": "Users can vote for their preffered candidate for various political offices",
        "operationId": "createvote",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "New Vote for an office",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Vote"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Vote registered successfully",
            "schema": {
              "$ref": "#/definitions/VoteModel"
            }
          },
          "400": {
            "description": "You have voted a candidate for this office"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "VoteModel": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "officeid": {
              "type": "string",
              "format": "uuid"
            },
            "candidateid": {
              "type": "string",
              "format": "uuid"
            },
            "userid": {
              "type": "string",
              "format": "uuid"
            },
            "created_date": {
              "type": "string",
              "format": "date-time"
            }
          }
        }
      }
    },
    "Vote": {
      "type": "object",
      "properties": {
        "candiate": {
          "type": "string",
          "format": "uuid"
        },
        "office": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
...
    "offices": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer"
        },
        "data": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "name": {
              "type": "string"
            },
            "type": {
              "type": "string"
            },
            "created_date": {
              "type": "string",
              "format": "date-time"
            },
            "modified_date": {
              "type": "string",
              "format": "date-time"
            }
          }
        }
      }
    },
...
    "parties": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer"
        },
        "data": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "name": {
              "type": "string"
            },
            "alias": {
              "type": "string"
            },
            "hqaddress": {
              "type": "string"
            },
            "logourl": {
              "type": "string"
            },
            "datecreated": {
              "type": "string",
              "format": "date-time"
            },
            "datemodified": {
              "type": "string",
              "format": "date-time"
            }
          }
        }
      }
    },
...
  }
}